+++ packages/argdown-cli/src/plugins/IncludePlugin.js
@@ -6,40 +6,31 @@ let path = require("path");
 import * as _ from "lodash";
 
 class IncludePlugin {
-    set config(config) {
-        let previousSettings = this.settings;
-        if (!previousSettings) {
-            previousSettings = {
-                regEx: /@include\(([^\)]+)\)/g
-            };
-        }
-        this.settings = _.defaultsDeep({}, config, previousSettings);
-    }
     constructor(config) {
         this.name = "IncludePlugin";
-        this.config = config;
+        this.defaults = _.defaultsDeep({}, config, {
+            regEx: /@include\(([^\)]+)\)/g
+        });
+    }
+    getSettings(request) {
+        if (!request.include) {
+            request.include = {};
+        }
+        return request.include;
+    }
+    prepare(request) {
+        _.defaultsDeep(this.getSettings(request), this.defaults);
     }
     async runAsync(request, response) {
-        if (request.include) {
-            this.config = request.include;
-        }
         if (!request.input || !request.inputPath) {
             return response;
         }
-        request.input = await this.replaceIncludesAsync(
-            request.inputPath,
-            request.input,
-            this.settings.regEx,
-            []
-        );
+        const settings = this.getSettings(request);
+        settings.regEx.lastIndex = 0;
+        request.input = await this.replaceIncludesAsync(request.inputPath, request.input, settings.regEx, []);
         return response;
     }
-    async replaceIncludesAsync(
-        currentFilePath,
-        str,
-        regEx,
-        filesAlreadyIncluded
-    ) {
+    async replaceIncludesAsync(currentFilePath, str, regEx, filesAlreadyIncluded) {
         let match = null;
         const directoryPath = path.dirname(currentFilePath);
         regEx.lastIndex = 0;
@@ -55,10 +46,7 @@ class IncludePlugin {
                 filesAlreadyIncluded.push(absoluteFilePath);
                 strToInclude = await readFile(absoluteFilePath, "utf8");
                 if (strToInclude == null) {
-                    strToInclude =
-                        "<!-- Include failed: File '" +
-                        absoluteFilePath +
-                        "' not found. -->\n";
+                    strToInclude = "<!-- Include failed: File '" + absoluteFilePath + "' not found. -->\n";
                 } else {
                     strToInclude = await this.replaceIncludesAsync(
                         absoluteFilePath,
@@ -68,10 +56,7 @@ class IncludePlugin {
                     );
                 }
             }
-            str =
-                str.substr(0, match.index) +
-                strToInclude +
-                str.substr(match.index + match[0].length);
+            str = str.substr(0, match.index) + strToInclude + str.substr(match.index + match[0].length);
             regEx.lastIndex = match.index + strToInclude.length;
         }
         return str;