+++ packages/argdown-cli/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.9.1
+
+### Minor Changes
+
+*   Update to argdown-parser 0.8.1 and argdown-map-maker 0.4.1 (see their Changelog for info about the new bug fixes)
+*   Added support for plugin's prepare method in app.runAsync
+*   Implemented new configuration pattern for plugins (for details see Readme of argdown-parser)
+
 ## 0.9.0
 
 ### Breaking Changes